Ready for refactoring

+++ BirdCounter/BirdCounter/Statistics.cs
@@ -1,12 +1,14 @@
 ﻿namespace BirdCounter
 {
-    public  class Statistics
+    public class Statistics
     {
         public int Min { get; set; }
         public int Max { get; set; }
-        public float Avarage { get
-            {
-                return Sum / Count;                
+        public float Avarage
+        {
+            get
+            {                
+                return (float)Sum / Count;
             }
         }
         public int Sum { get; set; }
@@ -17,20 +19,20 @@
         {
             Sum = 0;
             Count = 0;
-            
+
             Min = int.MaxValue;
             Max = int.MinValue;
-            
+
+
         }
 
-        public void CalculateStatistics (int number)
+        public void CalculateStatistics(int number)
         {
             Count++;
             Sum += number;
             Min = Math.Min(number, Min);
             Max = Math.Max(number, Max);
         }
-       
-     
+
     }
 }
